+++ src/orchestration/prefect_viadot/flows/exchange_rates_to_adls.py
@@ -1,12 +1,10 @@
-"""Flows for pulling data from Exchange rates API to Databricks."""
+"""Flows for pulling data from Exchange rates API to Azure Data Lake."""
 
 from datetime import datetime
 from typing import Literal
 
+from orchestration.prefect_viadot.tasks import df_to_adls, exchange_rates_to_df
 from prefect import flow
-
-from prefect_viadot.tasks import df_to_databricks, exchange_rates_to_df
-
 
 Currency = Literal[
     "USD", "EUR", "GBP", "CHF", "PLN", "DKK", "COP", "CZK", "SEK", "NOK", "ISK"
@@ -14,33 +12,29 @@ Currency = Literal[
 
 
 @flow(
-    name="extract--exchange-rates-api--databricks",
-    description="Extract data from Exchange Rates API and load it into Databricks.",
+    name="extract--exchange-rates-api--adls",
+    description="Extract data from Exchange Rates API and load it into Azure Data Lake.",
     retries=1,
     retry_delay_seconds=60,
 )
-def exchange_rates_to_databricks(  # noqa: PLR0913, PLR0917
-    databricks_table: str,
-    databricks_schema: str | None = None,
-    if_exists: Literal["replace", "skip", "fail"] = "fail",
+def exchange_rates_to_adls(  # noqa: PLR0913, PLR0917
+    adls_path: str,
+    overwrite: bool = False,
     currency: Currency = "USD",
     start_date: str = datetime.today().strftime("%Y-%m-%d"),
     end_date: str = datetime.today().strftime("%Y-%m-%d"),
     symbols: list[str] | None = None,
     exchange_rates_credentials_secret: str | None = None,
     exchange_rates_config_key: str | None = None,
-    databricks_credentials_secret: str | None = None,
-    databricks_config_key: str | None = None,
+    adls_credentials_secret: str | None = None,
+    adls_config_key: str | None = None,
 ) -> None:
-    """Download a DataFrame from ExchangeRates API and upload it to Databricks.
+    """Download a DataFrame from ExchangeRates API and upload it to Azure Data Lake.
 
     Args:
-        databricks_table (str): The name of the target table.
-        databricks_schema (str, optional): The name of the target schema.
-            Defaults to None.
-        if_exists (Literal["replace", "skip", "fail"], optional):
-            What to do if the table already exists.
-            One of "replace", "skip", and "fail". Defaults to "fail".
+        adls_path (str): The destination path.
+        overwrite (bool, optional): Whether to overwrite files in the lake.
+            Defaults to False.
         currency (Currency, optional): Base currency to which prices of searched
             currencies are related. Defaults to "USD".
         start_date (str, optional): Initial date for data search.
@@ -53,16 +47,15 @@ def exchange_rates_to_databricks(  # noqa: PLR0913, PLR0917
             exchange rates from base currency will be fetched.
             Defaults to
             ["USD","EUR","GBP","CHF","PLN","DKK","COP","CZK","SEK","NOK","ISK"].
-            Only ISO codes.
         exchange_rates_credentials_secret (str, optional): The name of the
             Azure Key Vault secret storing the exchange rates credentials.
             Defaults to None.
         exchange_rates_config_key (str, optional): The key in the viadot config holding
             relevant credentials. Defaults to None.
-        databricks_credentials_secret (str, optional): The name of the Azure Key Vault
-            secret storing relevant credentials. Defaults to None.
-        databricks_config_key (str, optional): The key in the viadot config holding
-            relevant credentials. Defaults to None.
+        adls_credentials_secret (str, optional): The name of the Azure Key Vault secret
+            storing the ADLS credentials. Defaults to None.
+        adls_config_key (str, optional): The key in the viadot config holding relevant
+            credentials. Defaults to None.
     """
     df = exchange_rates_to_df(
         currency=currency,
@@ -73,11 +66,17 @@ def exchange_rates_to_databricks(  # noqa: PLR0913, PLR0917
         symbols=symbols,
     )
 
-    return df_to_databricks(
+    return df_to_adls(
         df=df,
-        schema=databricks_schema,
-        table=databricks_table,
-        if_exists=if_exists,
-        credentials_secret=databricks_credentials_secret,
-        config_key=databricks_config_key,
+        path=adls_path,
+        credentials_secret=adls_credentials_secret,
+        config_key=adls_config_key,
+        overwrite=overwrite,
     )
+
+
+exchange_rates_to_adls(
+    adls_path="raw/supermetrics/tests/test_adls_prefect_viadot.parquet",
+    exchange_rates_config_key="exchange_rates_dev",
+    adls_config_key="adls_dev",
+)
