+++ src/orchestration/prefect_viadot/flows/sharepoint_to_redshift_spectrum.py
@@ -2,9 +2,8 @@
 
 from typing import Any, Literal
 
+from orchestration.prefect_viadot.tasks import df_to_redshift_spectrum, sharepoint_to_df
 from prefect import flow
-
-from prefect_viadot.tasks import df_to_redshift_spectrum, sharepoint_to_df
 
 
 @flow(
