+++ src/orchestration/prefect_viadot/flows/sharepoint_to_s3.py
@@ -2,10 +2,9 @@
 
 from typing import Any
 
+from orchestration.prefect_viadot.tasks import s3_upload_file, sharepoint_download_file
 from prefect import flow
 from viadot.sources.sharepoint import SharepointCredentials
-
-from prefect_viadot.tasks import s3_upload_file, sharepoint_download_file
 
 
 @flow(
