+++ src/orchestration/prefect_viadot/tasks/cloud_for_customers.py
@@ -3,10 +3,9 @@
 from typing import Any
 
 import pandas as pd
+from orchestration.prefect_viadot.utils import get_credentials
 from prefect import task
 from viadot.sources import CloudForCustomers
-
-from prefect_viadot.utils import get_credentials
 
 
 @task(retries=3, retry_delay_seconds=10, timeout_seconds=60 * 60)
