+++ src/orchestration/prefect_viadot/tasks/dbt.py
@@ -4,10 +4,9 @@ import logging
 import os
 from typing import Any
 
+from orchestration.prefect_viadot.utils import shell_run_command
 from prefect import task
 from prefect.logging import get_run_logger
-
-from prefect_viadot.utils import shell_run_command
 
 
 @task(retries=0, timeout_seconds=2 * 60 * 60)
