+++ src/orchestration/prefect_viadot/tasks/sap_rfc.py
@@ -7,11 +7,10 @@ import pandas as pd
 from prefect import task
 from prefect.logging import get_run_logger
 
-
 with contextlib.suppress(ImportError):
     from viadot.sources import SAPRFC, SAPRFCV2
-from prefect_viadot.exceptions import MissingSourceCredentialsError
-from prefect_viadot.utils import get_credentials
+from orchestration.prefect_viadot.exceptions import MissingSourceCredentialsError
+from orchestration.prefect_viadot.utils import get_credentials
 
 
 @task(retries=3, retry_delay_seconds=10, timeout_seconds=60 * 60 * 3)
